page web

+++ src/main/java/fr/cda/ecommerce/model/OrderProduct.java
@@ -38,7 +38,7 @@ public class OrderProduct {
         return "OrderProduct [product=" + product + ", quantity=" + quantity + "]";
     }
 
-    public Double getTotalPrice() {
+    public Integer getTotalPrice() {
         return getProduct().getPrice() * getQuantity();
     }
