+++ src/main/java/fr/cda/ecommerce/service/ProductService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ProductService {
     public abstract List<Product> getAllProducts();
-    public abstract Product getProductById(Long id);
+    public abstract Product getProductById(int id);
     public abstract Product save(Product product);
     public abstract boolean isProductAvailable(Product product, int quantity);
     public abstract void removeProduct(Product product, int quantity) throws StockException;
